# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Tres Service API Documentation",
        "description": "API documentation for Tres Service application",
        "contact": {
            "email": "sibomanaedouard974@gmail.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "API Server"
        }
    ],
    "paths": {
...
        "/api/v1/shared-with-me": {
            "get": {
                "tags": [
                    "Sharing & Links"
                ],
                "summary": "Get folders & files shared with the current user",
                "description": "Retrieves a list of files and folders shared with the authenticated user (via email share or internal share)",
                "operationId": "getSharedWithMe",
                "responses": {
                    "200": {
                        "description": "Shared items retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Shared items retrieved successfully"
                                        },
                                        "data": {
                                            "properties": {
                                                "shared_files": {
                                                    "type": "array",
                                                    "items": {
                                                        "$ref": "#/components/schemas/File"
                                                    }
                                                },
                                                "shared_folders": {
                                                    "type": "array",
                                                    "items": {
                                                        "$ref": "#/components/schemas/Folder"
                                                    }
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
            "File": {
                "title": "File",
                "description": "File model",
                "required": [
                    "id",
                    "user_id",
                    "folder_id",
                    "name",
                    "file_path",
                    "file_type",
                    "file_size",
                    "created_at",
                    "updated_at"
                ],
                "properties": {
                    "id": {
                        "description": "File ID",
                        "type": "integer",
                        "format": "int64",
                        "example": 1
                    },
                    "user_id": {
                        "description": "Owner User ID",
                        "type": "integer",
                        "format": "int64",
                        "example": 1
                    },
                    "folder_id": {
                        "description": "Parent Folder ID",
                        "type": "integer",
                        "format": "int64",
                        "example": 1
                    },
                    "name": {
                        "description": "File name",
                        "type": "string",
                        "example": "document.pdf"
                    },
                    "description": {
                        "description": "File description",
                        "type": "string",
                        "example": "A brief summary of the document.",
                        "nullable": true
                    },
                    "file_path": {
                        "description": "Internal path to the file storage",
                        "type": "string",
                        "example": "files/user_1/folder_1/document.pdf"
                    },
                    "file_type": {
                        "description": "MIME type of the file",
                        "type": "string",
                        "example": "application/pdf"
                    },
                    "file_size": {
                        "description": "File size in bytes",
                        "type": "integer",
                        "example": 1024000
                    },
                    "total_downloads": {
                        "description": "Total number of times the file has been downloaded",
                        "type": "integer",
                        "example": 50
                    },
                    "last_access_date": {
                        "description": "Last access timestamp",
                        "type": "string",
                        "format": "date-time",
                        "example": "2025-08-09T14:30:00Z",
                        "nullable": true
                    },
                    "created_at": {
                        "description": "Creation timestamp",
                        "type": "string",
                        "format": "date-time",
                        "example": "2025-08-09T12:00:00Z"
                    },
                    "updated_at": {
                        "description": "Last update timestamp",
                        "type": "string",
                        "format": "date-time",
                        "example": "2025-08-09T12:00:00Z"
                    },
                    "deleted_at": {
                        "description": "Deletion timestamp (for soft delete)",
                        "type": "string",
                        "format": "date-time",
                        "example": null,
                        "nullable": true
                    }
                },
                "type": "object"
            },
            "Folder": {
                "title": "Folder",
                "description": "Folder model",
                "required": [
                    "id",
                    "user_id",
                    "name",
                    "privacy",
                    "created_at",
                    "updated_at"
                ],
                "properties": {
                    "id": {
                        "description": "Folder ID",
                        "type": "integer",
                        "format": "int64",
                        "example": 1
                    },
                    "user_id": {
                        "description": "Owner User ID",
                        "type": "integer",
                        "format": "int64",
                        "example": 1
                    },
                    "parent_id": {
                        "description": "Parent folder ID",
                        "type": "integer",
                        "format": "int64",
                        "example": null,
                        "nullable": true
                    },
                    "name": {
                        "description": "Folder name",
                        "type": "string",
                        "example": "My Documents"
                    },
                    "privacy": {
                        "description": "Folder privacy",
                        "type": "string",
                        "enum": [
                            "public",
                            "private"
                        ],
                        "example": "private"
                    },
                    "password_protected": {
                        "description": "Is folder password protected?",
                        "type": "boolean",
                        "example": false
                    },
                    "allow_download": {
                        "description": "Allow download from this folder",
                        "type": "boolean",
                        "example": true
                    },
                    "watermark_images": {
                        "description": "Watermark images in this folder",
                        "type": "boolean",
                        "example": false
                    },
                    "created_at": {
                        "description": "Creation timestamp",
                        "type": "string",
                        "format": "date-time",
                        "example": "2025-08-09T12:00:00Z"
                    },
                    "updated_at": {
                        "description": "Last update timestamp",
                        "type": "string",
                        "format": "date-time",
                        "example": "2025-08-09T12:00:00Z"
                    },
                    "deleted_at": {
                        "description": "Deletion timestamp (for soft delete)",
                        "type": "string",
                        "format": "date-time",
                        "example": null,
                        "nullable": true
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            },
            "sanctum": {
                "type": "apiKey",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization",
                "in": "header"
            }
        }
    },
...
    "security": [
        {
            "sanctum": []
        }
    ]
}